lte/alt1250: Fix issue that caused assert when sending UDP packet
The reason for the issue is that the USRSOCK_REQUEST_SENDTO
handler does not read the send data from /dev/usrsock.
Since the handler does not read, the data remains in /dev/usrsock.
Therefore, the main loop of ALT1250 daemon detects the next
request from usersock and reads /dev/usrsock.
Then, the read data is different from the expected data,
so it asserts.