wireless/gs2200m: Implement handler of shutdown() request
shutdown() request has been implemented on the kernel side at
d3dd3496495d54de1bb9e8daf39828816e2c4f78 commit.
But no modification on gs2200m usrsock daemon side.
Because of it, when the daemon received the shutdown request,
the daemon is stacked.
This commit implements the handler of shutdown on it.